requirements.txt
python-telegram-bot==20.7
instaloader==4.10.1
python-dotenv==1.0.0
requests==2.31.0
Pillow==10.1.0
aiohttp==3.9.1
